Allow a file to be opened again with different access
HDS v4 allows a file to be opened for update even if
it has already been opened for read-only access.  See
rec1_open_file.c:

/* See if the file needs opening. This will be necessary if a new FCV slot  */
/* is being used or if the file is currently open for read-only access and  */
/* write access is now required.                                            */
      if ( _ok( hds_gl_status ) )
      {
         mustopen = *newslot ||
                    ( ( mode != 'R' ) &&
                      ( rec_ga_fcv[ *slot ].write == REC__NOIOCHAN ) );

But stock HDF5 reports an error if you try to do the same thing. See
H5F_open in file H5F.c. It turns out this causes problems for skyloop
(and possibly other apps), which make use of HDS's ability to re-open
files with different access. The change to HDF in this commit is the
easiest and least error-prone fix that I can think off for this.

It seems that HDF5 can in fact do what is required, but it has been
specifically disallowed by a check in H5F_open that reports an error if
an attempt is made to upgrade access from read-only to read-write.

This change involves defining a new file-open flag, called
H5F_ACC_FORCERW ("force read-write"), which can be supplied to H5F_open
to suppress the check described above, and thus allow the file to opened
with read-write access. HDS-V5 uses this flag when opening HDF5 files.
The behaviour of H5F_open us unchanged from stock HDF5 if this flag is
not supplied.