Update HDF5 from 1.8.13 to 1.10.4
Required because there is a serious bug in 1.8.13 that causes big JCMT DR
jobs (both POL2 and HARP) to abort with an error message:

"can't insert duplicate key"

According to https://github.com/h5py/h5py/issues/480, this is some resource
allocation issue within HDF5 that is only triggered for applications that
open files many times. It is apparently fixed in 1.8.14, but we may as
well catch up to the current version whilst we are at it.