Re-implement 27599d7bf "Allow a file to be opened again with different access"
The code that was in H5F.c in 1.8.13 is in H5Fint.c in 1.10.4. So the
modification to that code introduced by commit 27599d7bf on the master
branch needs to be reimplmented in H5Fint.c. Also, the value of the
FORCERW flag needs to be changed because HDF5 has used the old value for
other purposes.