Remove the starlink-specific FORCERW file-opening flag
The FORERW flag was indended to allow a file that had previously been
opened for read-only access to be re-opened for read-write access. It
worked in HDF version 1.8.13 but seems to have broken in version 1.10.4.
So this commit removes the FORCERW flag. Instead, HDS-V5 will be changed
to close the previously opened file, and then re-open it, preserving any
locators associated with the open file.